+++ si_mountainsort5/main.py
@@ -2,10 +2,8 @@
 
 from dendro.sdk import App, ProcessorBase
 from pydantic import Field
-from common.models import (
-    MountainSort5SortingContext,
-    PipelineFullContext
-)
+from common.models import PipelineFullContext
+from common.models_sorting import MountainSort5SortingContext
 from common.processor_pipeline import run_pipeline
